// YOUCOM.ReserVook/src/app/features/web/master/paymentconvert/model/paymentconvert.model.ts
export class PaymentConvertInfo {
  companyNo: string;
  scCd: string;
  scSiteCd: string;
  scPaymentOpts: string;
  denominationCode: string;
  updateCnt: number;
  programId: string;
  createClerk: string;
  createMachineNo: string;
  createMachine: string;
  createDatetime: string;
  updateClerk: string;
  updateMachineNo: string;
  updateMachine: string;
  updateDatetime: string;
  status: string;

  cdName: string;
  siteCdName: string;
  scPaymentOptsName: string;
  denomination: string;
}